feat: update financial reporting mocks with new approval IDs
