+++ src/mocks/financialReporting/Approvals.mock.ts
@@ -1,0 +1,114 @@
+export const approval_by_credit_request_Mock = [
+  {
+    approval_id: "05af9ef8-0fce-4409-9114-311eb7559a61",
+    credit_request_id: "999990",
+    approver_id: "4156416",
+    approver_name: "Pedro Pablo Iregui Gerrero",
+    concept: "Aprobado",
+    error: false,
+  },
+  {
+    approval_id: "a6299554-84a3-4db4-b0fb-a7ed170732e6",
+    credit_request_id: "999990",
+    approver_id: "8945612",
+    approver_name: "Carlos Alberto Combita",
+    concept: "Rechazado",
+    error: false,
+  },
+  {
+    approval_id: "a87c3922-0834-471f-9309-6976883b6671",
+    credit_request_id: "999990",
+    approver_id: "7589456",
+    approver_name: "Jaime Alberto Linares Guacaneme",
+    concept: "Rechazado",
+    error: false,
+  },
+  {
+    approval_id: "3bc6faa3-9448-4e27-a21c-c33da5b6fc95",
+    credit_request_id: "999990",
+    approver_id: "1597538",
+    approver_name: "Miguel Angel Fuentes",
+    concept: "Pendiente",
+    error: true,
+  },
+  {
+    approval_id: "b4512600-f758-4a9e-9fdf-40266ddc3cbc",
+    credit_request_id: "999990",
+    approver_id: "3571598",
+    approver_name: "Cesar Augusto Corredor",
+    concept: "Aprobado",
+    error: false,
+  },
+  {
+    approval_id: "9fbf0b04-2294-440c-a6ff-9ba08594a0c8",
+    credit_request_id: "999990",
+    approver_id: "8523697",
+    approver_name: "Paula Andrea Betancurt",
+    concept: "Rechazado",
+    error: false,
+  },
+  {
+    approval_id: "44683722-e703-4c85-a59a-9710d31edc77",
+    credit_request_id: "999990",
+    approver_id: "9517536",
+    approver_name: "Jaime Alejandro Vargas",
+    concept: "Pendiente",
+    error: true,
+  },
+  {
+    approval_id: "05af9ef8-0fce-4409-9114-311eb7559a61",
+    credit_request_id: "999991",
+    approver_id: "4156416",
+    approver_name: "Pedro Pablo Iregui Gerrero",
+    concept: "Aprobado",
+    error: false,
+  },
+  {
+    approval_id: "a6299554-84a3-4db4-b0fb-a7ed170732e6",
+    credit_request_id: "999991",
+    approver_id: "8945612",
+    approver_name: "Carlos Alberto Combita",
+    concept: "Rechazado",
+    error: false,
+  },
+  {
+    approval_id: "a87c3922-0834-471f-9309-6976883b6671",
+    credit_request_id: "999991",
+    approver_id: "7589456",
+    approver_name: "Jaime Alberto Linares Guacaneme",
+    concept: "Rechazado",
+    error: false,
+  },
+  {
+    approval_id: "3bc6faa3-9448-4e27-a21c-c33da5b6fc95",
+    credit_request_id: "999992",
+    approver_id: "1597538",
+    approver_name: "Miguel Angel Fuentes",
+    concept: "Pendiente",
+    error: true,
+  },
+  {
+    approval_id: "b4512600-f758-4a9e-9fdf-40266ddc3cbc",
+    credit_request_id: "999992",
+    approver_id: "3571598",
+    approver_name: "Cesar Augusto Corredor",
+    concept: "Aprobado",
+    error: false,
+  },
+  {
+    approval_id: "9fbf0b04-2294-440c-a6ff-9ba08594a0c8",
+    credit_request_id: "999993",
+    approver_id: "8523697",
+    approver_name: "Paula Andrea Betancurt",
+    concept: "Rechazado",
+    error: false,
+  },
+  {
+    approval_id: "44683722-e703-4c85-a59a-9710d31edc77",
+    credit_request_id: "999993",
+    approver_id: "9517536",
+    approver_name: "Jaime Alejandro Vargas",
+    concept: "Pendiente",
+    error: true,
+  },
+];
