--- src/mocks/financialReporting/Approvals.ts
@@ -1,58 +1,0 @@
-export const approval_by_credit_request_Mock = [
-  {
-    approval_id: "05af9ef8-0fce-4409-9114-311eb7559a61",
-    credit_request_id: "e6b54a96-52bf-4b52-b8ce-b5870fe7a32d",
-    approver_id: "4156416",
-    approver_name: "Pedro Pablo Iregui Gerrero",
-    concept: "Aprobado",
-    error: false,
-  },
-  {
-    approval_id: "a6299554-84a3-4db4-b0fb-a7ed170732e6",
-    credit_request_id: "b3c45d22-1c0d-4d9a-ae3e-973ad7fbc84d",
-    approver_id: "8945612",
-    approver_name: "Carlos Alberto Combita",
-    concept: "Rechazado",
-    error: false,
-  },
-  {
-    approval_id: "a87c3922-0834-471f-9309-6976883b6671",
-    credit_request_id: "a2d3d634-0fda-4972-a412-56b43ed33b34",
-    approver_id: "7589456",
-    approver_name: "Jaime Alberto Linares Guacaneme",
-    concept: "Rechazado",
-    error: false,
-  },
-  {
-    approval_id: "3bc6faa3-9448-4e27-a21c-c33da5b6fc95",
-    credit_request_id: "f1a2b3c4-56d7-48b2-9f80-a6c5a7e3a2b1",
-    approver_id: "1597538",
-    approver_name: "Miguel Angel Fuentes",
-    concept: "Pendiente",
-    error: true,
-  },
-  {
-    approval_id: "b4512600-f758-4a9e-9fdf-40266ddc3cbc",
-    credit_request_id: "b9c8a7d6-5432-4a65-9b9c-d8f7a6e4a3b2",
-    approver_id: "3571598",
-    approver_name: "Cesar Augusto Corredor",
-    concept: "Aprobado",
-    error: false,
-  },
-  {
-    approval_id: "9fbf0b04-2294-440c-a6ff-9ba08594a0c8",
-    credit_request_id: "c4b3a2d1-9087-4f6d-873e-5a4b3c6d7a8f",
-    approver_id: "8523697",
-    approver_name: "Paula Andrea Betancurt",
-    concept: "Rechazado",
-    error: false,
-  },
-  {
-    approval_id: "44683722-e703-4c85-a59a-9710d31edc77",
-    credit_request_id: "d5e4f3a2-6789-4b6a-8d7e-9f5a6b4c3a2b",
-    approver_id: "9517536",
-    approver_name: "Jaime Alejandro Vargas",
-    concept: "Pendiente",
-    error: true,
-  },
-];
